@@ -43,8 +43,8 @@
     "babel-preset-env": "^1.6.0",
     "babel-preset-flow": "^6.23.0",
     "chalk": "^2.1.0",
-    "eslint": "^4.7.1",
-    "eslint-config-airbnb": "^15.1.0",
+    "eslint": "^4.10.0",
+    "eslint-config-airbnb": "^16.1.0",
     "eslint-config-prettier": "^2.4.0",
     "eslint-plugin-flowtype": "^2.35.1",
     "eslint-plugin-import": "^2.7.0",
