@@ -2,12 +2,20 @@
 import type { Context } from 'koa';
 import { type Log } from 'neo-blockchain-node-core';
 
-// $FlowFixMe
-import { performance } from 'perf_hooks'; // eslint-disable-line
+import { getLog, simpleMiddleware } from './common';
 
-export default function({ log }: {| log: Log |}) {
-  return async (ctx: Context, next: () => Promise<void>) => {
-    const startTime = performance.now();
+const nowSeconds = () => Date.now() / 1000;
+
+const getLatencySeconds = (startTimeSeconds: number) => {
+  const latencySeconds = nowSeconds() - startTimeSeconds;
+  return Math.round(latencySeconds * 1000) / 1000;
+};
+
+export default simpleMiddleware(
+  'logger',
+  async (ctx: Context, next: () => Promise<void>) => {
+    const startTimeSeconds = nowSeconds();
+    const log = getLog(ctx);
     try {
       await next();
     } catch (error) {
@@ -15,7 +23,7 @@ export default function({ log }: {| log: Log |}) {
         log({
           event: 'REQUEST_ERROR',
           error,
-          durationMS: performance.now() - startTime,
+          httpRequest: { latency: getLatencySeconds(startTimeSeconds) },
         });
       }
 
@@ -23,17 +31,25 @@ export default function({ log }: {| log: Log |}) {
     } finally {
       log({
         event: 'REQUEST',
-        durationMS: performance.now() - startTime,
+        httpRequest: { latency: getLatencySeconds(startTimeSeconds) },
       });
     }
-  };
-}
+  },
+);
 
-export function onError({ log }: {| log: Log |}) {
-  return (error: Error) => {
-    log({
-      event: 'UNEXPECTED_REQUEST_ERROR',
-      error,
-    });
-  };
-}
+export const onError = ({ log: logIn }: {| log: Log |}) => (
+  error: Error,
+  ctx?: Context,
+) => {
+  let log = logIn;
+  if (ctx != null) {
+    try {
+      log = getLog(ctx);
+      // eslint-disable-next-line
+    } catch (error) {
+      // eslint-disable-next-line
+    }
+  }
+
+  log({ event: 'UNEXPECTED_REQUEST_ERROR', error });
+};
